+++ CHANGELOG.md
@@ -6,12 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Changed
-- Simplified mutex and stream server lock timeout handling.
+- Simplified mutex, stream server and buffer lock timeout handling.
 
 ### Fixed
 - Event handler memory leak.
 - Missing lock in stream read buffer flushing.
 - Unsynchronized stream server task state flags.
+- Silent buffer allocation failure.
 
 ## [1.2.2] - 2024-09-10
 ### Changed